优化了牛耕田法算法中的tsp问题，整体规划时间可以缩短三分之一左右 增加加了缓存机制，建图、修改禁区后会规划一次，之后遍会直接获取此处生成的全覆盖路径 因为增加了缓存，所以每次获得的规划路径就一样，自然不能根据当前机器人位置动态的规划

+++ include/model/ParamVo.h
@@ -14,7 +14,7 @@ class ParamVo {
 private:
     int tof;
     bool silver;
-    int speed;
+    float speed;
 public:
     ParamVo();
 
@@ -28,9 +28,9 @@ public:
 
     void setSilver(bool silver);
 
-    int getSpeed() const;
+    float getSpeed() const;
 
-    void setSpeed(int speed);
+    void setSpeed(float speed);
 
     friend void to_json(json &j, const ParamVo &b) {
         j = json{
